backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
